+++ src/components/Layout/Footer/FooterLanguageSwitch.astro
@@ -1,5 +1,6 @@
 ---
-import { getLanguageFromURL, LOCALE_NAMES } from "@i18n/locales";
+import { LOCALE_NAMES } from "@i18n/locales";
+import { getLanguageFromURL } from "@i18n/utils";
 import LanguageSwitch from "../LanguageSwitch";
 import { Icon } from "@adjust/components";
