+++ src/components/Layout/Header/MainHeader.astro
@@ -1,7 +1,7 @@
 ---
 import { Icon } from "@adjust/components";
 
-import { getLanguageFromURL } from "@i18n/locales";
+import { getLanguageFromURL } from "@i18n/utils";
 import AudienceDropdown from "../AudienceDropdown";
 import { HELP_CENTER_LINK } from "src/consts";
