+++ src/components/Layout/Header/MobileHeader.astro
@@ -1,6 +1,6 @@
 ---
 import IconSearch from "@components/Icons/astro/IconSearch.astro";
-import { getLanguageFromURL } from "@i18n/locales";
+import { getLanguageFromURL } from "@i18n/utils";
 
 import MenuToggle from "./SidebarToggle";
