+++ src/components/Layout/LeftSidebar/Navigation.astro
@@ -1,7 +1,7 @@
 ---
 import classNames from "classnames";
 
-import { getLanguageFromURL } from "@i18n/locales";
+import { getLanguageFromURL } from "@i18n/utils";
 import NavigationItem from "./NavigationItem";
 import { getCurrentPage } from "@utils/helpers/navigation/getCurrentPage";
