+++ src/i18n/locales.ts
@@ -7,13 +7,5 @@ export const LOCALE_NAMES: Locales = {
   ko: "한국어",
 };
 
-export const langPathRegex = /\/([a-z]{2}-?[A-Z]{0,2})(\/|$)/;
-export const langParamRegex = /(^|\/)([a-z]{2}-?[A-Z]{0,2})(\/|$)/;
-
 export const KNOWN_LANGUAGE_CODES = Object.keys(LOCALE_NAMES);
-
-export const getLanguageFromURL = (pathname: string) => {
-  const langCodeMatch = pathname.match(langPathRegex);
-  const langCode = langCodeMatch ? langCodeMatch[1] : "en";
-  return langCode as keyof Locales;
-};
+export const defaultLang = "en";
