+++ src/layouts/MainLayout.astro
@@ -12,7 +12,8 @@ import { getNavigationEntries } from "src/utils/helpers/navigation/getNavigation
 import SidebarHeader from "@components/Layout/LeftSidebar/SidebarHeader/SidebarHeader.astro";
 import SidebarSearch from "@components/Layout/LeftSidebar/SidebarSearch/SidebarSearch.astro";
 import LanguageSwitch from "@components/Layout/LanguageSwitch";
-import { LOCALE_NAMES, getLanguageFromURL } from "@i18n/locales";
+import { LOCALE_NAMES } from "@i18n/locales";
+import { getLanguageFromURL } from "@i18n/utils";
 import MobileHeader from "@components/Layout/Header/MobileHeader.astro";
 import Footer from "@components/Layout/Footer/FooterContent.astro";
 import BodyNoScript from "@components/Layout/GoogleTagManager/BodyNoScript.astro";
